DROTH-3076 Moved Velho Lane Apis to own Api
