+++ digiroad2-api-oth/src/main/scala/fi/liikennevirasto/digiroad2/IntegrationApi.scala
@@ -28,16 +28,11 @@ import org.slf4j.LoggerFactory
 import scala.util.Try
 
 class IntegrationApi(val massTransitStopService: MassTransitStopService, implicit val swagger: Swagger) extends ScalatraServlet with JacksonJsonSupport with SwaggerSupport {
-  lazy val vkmClient = new VKMClient
-  lazy val polygonTools = new PolygonTools
   val logger = LoggerFactory.getLogger(getClass)
   protected val applicationDescription = "Integration API "
   protected implicit val jsonFormats: Formats = DefaultFormats
 
   case class AssetTimeStamps(created: Modification, modified: Modification) extends TimeStamps
-  case class VelhoRoad(roadNumber: Long, roadParts: Seq[RoadPart])
-  case class RoadPart(roadNumber: Option[Long], roadPartNumber: Long, lanes: Seq[VelhoLane])
-  case class VelhoLane(laneCode: Int, startAddressM: Long, endAddressM: Long)
 
   def extractModificationTime(timeStamps: TimeStamps): (String, String) = {
     "muokattu_viimeksi" ->
@@ -166,130 +161,6 @@ class IntegrationApi(val massTransitStopService: MassTransitStopService, implici
   private def getMassTransitStopsByMunicipality(municipalityNumber: Int): Iterable[PersistedMassTransitStop] = {
     massTransitStopService.getByMunicipality(municipalityNumber)
   }
-
-  def lanesInRoadAddressSpaceToApi(roadNumber: Long, track: Track, startRoadPartNumber: Long, endRoadPartNumber: Long,
-                                   startAddrM: Long, endAddrM: Long): Seq[Map[String, Any]] = {
-    val roadAddresses = roadAddressService.getAllByRoadNumber(roadNumber).filter(_.track == track)
-    val roadPartMaxLengths = roadAddresses.groupBy(_.roadPartNumber).mapValues(_.maxBy(_.endAddrMValue)).values
-    val roadPartRange = startRoadPartNumber to endRoadPartNumber
-    val filteredMaxLengths = roadPartMaxLengths.filter(address => roadPartRange contains address.roadPartNumber)
-
-    val roadAddressesToTransform = filteredMaxLengths.flatMap(roadPart => {
-      val roadPartNumber = roadPart.roadPartNumber
-
-      val maxLength = if (roadPartNumber == endRoadPartNumber) endAddrM
-      else roadPart.endAddrMValue
-
-      val startLength = if (roadPartNumber == startRoadPartNumber) startAddrM
-      else roadPart.startAddrMValue
-
-      val transFormInterval = 50
-      val transformsForRoadPart = ((maxLength - startLength) / transFormInterval + 1).toInt
-
-      val roadAddresses = for (i <- 0 to transformsForRoadPart)
-        yield (i.toString, RoadAddress(None, roadNumber.toInt, roadPartNumber.toInt, track, (startLength + (i * transFormInterval)).toInt))
-      roadAddresses
-    }).toMap
-
-    val coordinatesAndIdentifiers = vkmClient.addressToCoordsMassQuery(roadAddressesToTransform).toMap
-    val polygon = polygonTools.createPolygonFromCoordinates(coordinatesAndIdentifiers)
-
-    val roadlinks = getRoadLinksAndChangesFromVVHWithPolygon(polygon)._1
-    val (roadLinksWithViiteAddress, noRoadAddress) = roadAddressService.roadLinkWithRoadAddress(roadlinks).partition(_.attributes.contains("VIITE_ROAD_NUMBER"))
-    val roadLinksWithTempAddress = roadAddressService.roadLinkWithRoadAddressTemp(noRoadAddress).filter(_.attributes.contains("TEMP_ROAD_NUMBER"))
-    val roadLinksWithRoadAddress = roadLinksWithViiteAddress ++ roadLinksWithTempAddress
-
-    val roadLinksWithConsistentRoadAddress = roadLinksWithConsistentAddress(roadLinksWithRoadAddress)
-
-    val correctLinks = roadLinksWithConsistentRoadAddress.filter(roadLink => roadLink.attributes("VIITE_ROAD_NUMBER") == roadNumber
-      && (roadPartRange contains roadLink.attributes("VIITE_ROAD_PART_NUMBER")) && roadLink.attributes("VIITE_TRACK") == track.value)
-
-    val lanesOnRoadLinks = laneService.getLanesByRoadLinks(correctLinks)
-    val (lanesWithViiteAddress, lanesWithoutRoadAddress) = roadAddressService.laneWithRoadAddress(Seq(lanesOnRoadLinks))
-      .flatten.partition(_.attributes.contains("VIITE_ROAD_NUMBER"))
-    val lanesWithTempAddress = roadAddressService.experimentalLaneWithRoadAddress(Seq(lanesWithoutRoadAddress)).flatten
-      .filter(_.attributes.contains("TEMP_ROAD_NUMBER"))
-    val lanesWithRoadAddress = lanesWithViiteAddress ++ lanesWithTempAddress
-
-    val lanesWithNormalRoadAddress = lanesWithConsistentRoadAddress(lanesWithRoadAddress)
-    val twoDigitLanes = laneService.pieceWiseLanesToTwoDigitWithMassQuery(lanesWithNormalRoadAddress).flatten
-
-    val velhoLanes = lanesToVelhoLanes(twoDigitLanes, correctLinks.groupBy(_.linkId).mapValues(_.head))
-    velhoLanes.map { velhoLane =>
-      Map("roadNumber" -> velhoLane.roadNumber,
-        "roadParts" -> velhoLane.roadParts
-      )
-    }
-  }
-
-  def lanesToVelhoLanes(twoDigitLanes: Seq[PieceWiseLane], roadLinks: Map[Long, RoadLink]): Seq[VelhoRoad] = {
-    val roadNumbers = twoDigitLanes.map(_.attributes("ROAD_NUMBER")).distinct.asInstanceOf[Seq[Long]]
-    val lanesGroupedByRoadNumber = twoDigitLanes.groupBy(_.attributes("ROAD_NUMBER")).values
-    val lanesGroupedByRoadNumberAndPartNumber = lanesGroupedByRoadNumber.flatMap(_.groupBy(_.attributes("ROAD_PART_NUMBER")).values)
-
-    val allRoadParts = lanesGroupedByRoadNumberAndPartNumber.map(lanes => {
-      val roadPartNumber = lanes.head.attributes("ROAD_PART_NUMBER").asInstanceOf[Long]
-      val roadNumber = Some(lanes.head.attributes("ROAD_NUMBER").asInstanceOf[Long])
-      val lanesGroupedByLaneCode = lanes.groupBy(_.laneAttributes.find(_.publicId == "lane_code")).values
-
-      val lanesWithContinuing = lanesGroupedByLaneCode.map(laneGroup => {
-        laneGroup.map(lane => {
-          val roadIdentifier = roadLinks(lane.linkId).roadIdentifier
-          val continuingLanes = LanePartitioner.getContinuingWithIdentifier(lane, roadIdentifier, laneGroup, roadLinks, true)
-          LaneWithContinuingLanes(lane, continuingLanes)
-        })
-      }).toSeq
-
-
-      val connectedGroups = lanesWithContinuing.flatMap(laneGroup => {
-        val startingLanes = getStartingLanes(laneGroup)
-        val connectedLanes = startingLanes.map(startingLane => {
-          getConnectedLanes(Seq(startingLane), laneGroup).sortBy(_.lane.id)
-        }).distinct
-        connectedLanes.map(_.map(_.lane))
-      })
-
-      val velhoLanes = connectedGroups.map(group => {
-        val laneCode = group.head.laneAttributes.find(_.publicId == "lane_code").get.values.head.value.asInstanceOf[Int]
-        val startLane = group.minBy(_.attributes("START_ADDR").asInstanceOf[Long])
-        val startAddressM = startLane.attributes("START_ADDR").asInstanceOf[Long] + startLane.startMeasure.asInstanceOf[Long]
-
-        val endLane = group.maxBy(_.attributes("END_ADDR").asInstanceOf[Long])
-        val endLaneLength = (roadLinks(endLane.linkId).length - endLane.endMeasure).asInstanceOf[Long]
-        val endAddressM = endLane.attributes("END_ADDR").asInstanceOf[Long] - endLaneLength
-
-        VelhoLane(laneCode, startAddressM, endAddressM)
-      })
-
-      RoadPart(roadNumber, roadPartNumber, velhoLanes)
-    }).toSeq
-
-    roadNumbers.map(roadNumber => {
-      val roadParts = allRoadParts.filter(_.roadNumber.get == roadNumber)
-      val roadPartsWithOutRoadNumber = roadParts.map(_.copy(roadNumber = None))
-      VelhoRoad(roadNumber, roadPartsWithOutRoadNumber)
-    })
-  }
-
-  def lanesToApi(municipalityNumber: Int): Seq[Map[String, Any]] = {
-    val roadLinksInMunicipality = roadLinkService.getRoadLinksFromVVH(municipalityNumber)
-    val lanes = laneService.getLanesByRoadLinks(roadLinksInMunicipality)
-
-    val (lanesWithViiteAddress, noRoadAddress) = roadAddressService.laneWithRoadAddress(Seq(lanes)).flatten.partition(_.attributes.contains("VIITE_ROAD_NUMBER"))
-    val lanesWithTempAddress =  roadAddressService.experimentalLaneWithRoadAddress(Seq(noRoadAddress)).flatten.filter(_.attributes.contains("TEMP_ROAD_NUMBER"))
-    val lanesWithRoadAddress = lanesWithViiteAddress ++ lanesWithTempAddress
-    val lanesWithNormalRoadAddress = lanesWithConsistentRoadAddress(lanesWithRoadAddress)
-
-    val twoDigitLanes = laneService.pieceWiseLanesToTwoDigitWithMassQuery(lanesWithNormalRoadAddress).flatten
-    val velhoLanes = lanesToVelhoLanes(twoDigitLanes, roadLinksInMunicipality.groupBy(_.linkId).mapValues(_.head))
-
-    velhoLanes.map { velhoLane =>
-      Map("roadNumber" -> velhoLane.roadNumber,
-          "roadParts" -> velhoLane.roadParts
-      )
-    }
-  }
-
 
   def speedLimitsToApi(speedLimits: Seq[SpeedLimit]): Seq[Map[String, Any]] = {
     speedLimits.map { speedLimit =>
@@ -1046,38 +917,6 @@ class IntegrationApi(val massTransitStopService: MassTransitStopService, implici
     }
   }
 
-  //Description of Api entry point to get all lanes in specific road address range
-  val getLanesInRoadAddressRange =
-    (apiOperation[Long]("getLanesInRoadAddressRange")
-      .parameters(
-        queryParam[Int]("roadNumber").description("Municipality Code where we will execute the search by specific asset type"),
-        queryParam[Int]("track").description("Municipality Code where we will execute the search by specific asset type"),
-        queryParam[Int]("startRoadPartNumber").description("Municipality Code where we will execute the search by specific asset type"),
-        queryParam[Int]("startAddrM").description("Municipality Code where we will execute the search by specific asset type"),
-        queryParam[Int]("endRoadPartNumber").description("Municipality Code where we will execute the search by specific asset type"),
-        queryParam[Int]("endAddrM").description("Municipality Code where we will execute the search by specific asset type"),
-        pathParam[String]("lanes_in_range").description("Asset type name to get all assets")
-      )
-      tags "Integration API (Kalpa API)"
-      summary "Get lanes in given road address range"
-      authorizations "Contact your service provider for more information"
-      description "Example URL: /externalApi/integration/lanes_in_range?road_number=9?track=1?star_part=208?start_addr=8500?end_part=208?end_addr=9000"
-      )
-
-  get("/lanes_in_range", operation(getLanesInRoadAddressRange)) {
-    contentType = formats("json")+ "; charset=utf-8"
-    val roadNumber = params("road_number").toLong
-    val track = Track(params("track").toInt)
-    val startRoadPartNumber = params("star_part").toLong
-    val startAddrM = params("start_addr").toLong
-    val endRoadPartNumber = params("end_part").toLong
-    val endAddrM = params("end_addr").toLong
-
-    lanesInRoadAddressSpaceToApi(roadNumber, track, startRoadPartNumber,endRoadPartNumber ,startAddrM , endAddrM)
-
-
-}
-
   //Description of Api entry point to get all assets by asset type and municipality
   val getAssetsByTypeMunicipality =
     (apiOperation[Long]("getAssetsByTypeMunicipality")
@@ -1140,7 +979,6 @@ class IntegrationApi(val massTransitStopService: MassTransitStopService, implici
         case "road_works_asset" => roadWorksToApi(municipalityNumber)
         case "parking_prohibitions" => parkingProhibitionsToApi(municipalityNumber)
         case "cycling_and_walking" => linearAssetsToApi(CyclingAndWalking.typeId, municipalityNumber)
-        case "lanes" => lanesToApi(municipalityNumber)
         case _ => BadRequest("Invalid asset type")
       }
     } getOrElse {
